chore: flatten highlight services into utils
